Split out contract structs to own file
